将一些数据从布局中拿出
