+++ src/layout/Layout.vue
@@ -1,5 +1,5 @@
 <template>
-  <a-layout id="components-layout-demo-top-side-2">
+  <a-layout>
     <a-layout-header class="header" style="padding: 0px">
       <div class="logo"/>
       <a-menu
@@ -8,29 +8,9 @@
           mode="horizontal"
           theme="dark"
       >
-        <a-menu-item key="/win-rate">
-          <router-link to="/win-rate">
-            阵容胜率排行
-          </router-link>
-        </a-menu-item>
-        <a-menu-item key="/rank">
-          <router-link to="/rank">
-            排行榜
-          </router-link>
-        </a-menu-item>
-        <a-menu-item key="/recommend">
-          <router-link to="/recommend">
-            模拟翻牌
-          </router-link>
-        </a-menu-item>
-        <a-menu-item key="/more-tools">
-          <router-link to="/more-tools">
-            攻略&工具导航
-          </router-link>
-        </a-menu-item>
-        <a-menu-item key="/tutorial">
-          <router-link to="/tutorial">
-            食用指南
+        <a-menu-item v-for="(value,name) in routes" :key="'/'+name">
+          <router-link :to="'/' + name">
+            {{ value.name }}
           </router-link>
         </a-menu-item>
       </a-menu>
@@ -49,8 +29,7 @@
       <a-layout-content
           :style="{ background: '#fff', padding: '24px', margin: 0, minHeight: '280px' }"
       >
-        <router-view></router-view>
-
+        <router-view/>
       </a-layout-content>
     </a-layout>
   </a-layout>
@@ -58,8 +37,7 @@
 
 <script>
 import config from "@/config";
-import battleTable from "../../data/data.json";
-import moment from "moment";
+import {updateTips} from "@/components/UpdateTips";
 
 export default {
   name: "Layout",
@@ -69,25 +47,7 @@ export default {
     }
   },
   mounted() {
-    if (this.$cookies.get('last_updated') !== config.update) {
-      const handleOk = () => {
-        this.$cookies.set('last_updated', config.update, moment().add(1000, 'days').toDate())
-      }
-      const h = this.$createElement;
-      this.$info({
-        width: '36%',
-        maskClosable: true,
-        title: 'Update',
-        content: h('div', {}, [
-          h('h3', config.update),
-          h('ul', {}, [
-            h('li', '数据更新：更新了' + config.range + '式神排行榜及' + battleTable.data.length + '场斗技数据。')
-          ]),
-        ]),
-        onOk: handleOk,
-        onCancel: handleOk,
-      });
-    }
+    updateTips();
   },
   methods: {
     refreshTitle() {
@@ -105,23 +65,10 @@ export default {
   data() {
     return {
       collapsed: false,
-      title: 'title',
+      title: '',
       activeRoute: '',
-      subTitle: 'sub-title',
-      // routes: [
-      //   {
-      //     path: 'index',
-      //     breadcrumbName: 'First-level Menu',
-      //   },
-      //   {
-      //     path: 'first',
-      //     breadcrumbName: 'Second-level Menu',
-      //   },
-      //   {
-      //     path: 'second',
-      //     breadcrumbName: 'Third-level Menu',
-      //   },
-      // ],
+      subTitle: '',
+      routes: {...config.routes},
     };
   },
 }
